PointRanking 10 위 까지가 아니라 페이지로 제공

+++ app/mvc-api/src/main/kotlin/com/anycommunity/mvcapi/point/response/PointRankingResponse.kt
@@ -8,7 +8,7 @@ import com.anycommunity.usecase.point.query.usecase.GetPointRankingResult
 @Schema(description = "포인트 랭킹 유저 조회 응답")
 data class PointRankingResponse(
     @field:Schema(description = "랭킹", example = "1")
-    val ranking: Int,
+    val ranking: Long,
     @field:Schema(description = "유저 id", example = "1")
     val userId: Long,
     @field:Schema(description = "사용자 이름", example = "ukjjang")
